+++ NotEnoughPotions/Assets/Data/Player Inventory.asset
@@ -9,7 +9,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 716a5c72f6e9b2b42ae04f9edff24246, type: 3}
+  m_Script: {fileID: 11500000, guid: d4f7a3008261a0846926f42b3fdef89d, type: 3}
   m_Name: Player Inventory
   m_EditorClassIdentifier: 
   Container: []
