+++ NotEnoughPotions/Assets/Scripts/Customer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: db12765dfc05ffc43b74fe6d41ffc187
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
