--- NotEnoughPotions/Assets/Scripts/InventoryObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 716a5c72f6e9b2b42ae04f9edff24246
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
