+++ NotEnoughPotions/Assets/Scripts/Potions/MakePotion.cs.meta
@@ -3,7 +3,8 @@ guid: 66af6152877393d4b859cc47214fcda9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
-  defaultReferences: []
+  defaultReferences:
+  - inventory: {fileID: 11400000, guid: 4c77335da6be5454e92ec0750af28f74, type: 2}
   executionOrder: 0
   icon: {instanceID: 0}
   userData:
