+++ NotEnoughPotions/Assets/Scripts/Scene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 52cb2ef23aaf7ce4abf3751e5cc14b39
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
